+++ src/lidia/lidia.py
@@ -1,10 +1,15 @@
 import argparse
+import json
 from multiprocessing import Process, Queue
+import tomli
+import re
 from typing import Dict
+import yaml
 
 from . import __version__
 from .server import run_server
 from .mytypes import RunFn, SetupFn
+from .config import Config
 
 from .sources import demo, rpctask, approach
 
@@ -17,6 +22,14 @@ def main():
         epilog='you can also see help for a specific source: "lidia <src> --help"')
     parser.add_argument('-V', '--version', action='version', version=f'{parser.prog} {__version__}',
                         help='display package version')
+    parser.add_argument('-v', '--verbose', action='count', dest='verbosity',
+                        help='increase amount of printed information, can be added multiple times', default=0)
+    parser.add_argument('-q', '--quiet', action='count',
+                        help='decrease amount of printed information', default=0)
+    parser.add_argument('-c', '--config', type=str,
+                        help='config files separated with comma, in order of applying')
+    parser.add_argument('-C', '--config-keys', type=str,
+                        help='properties as comma separated dotted keys')
     parser.add_argument('-H', '--http-host', type=str,
                         help='hosts to accept for web page', default='0.0.0.0')
     parser.add_argument('-P', '--http-port', type=int,
@@ -31,22 +44,55 @@ def main():
         sources[name] = run_function
 
     args = parser.parse_args()
+    args.verbosity -= args.quiet
+    if args.verbosity >= 1:
+        # mimic = specifier from python 3.8 to maintain 3.7 compatibility
+        print(f'args={args}')
+
+    config = Config()
+    if args.config is not None:
+        for config_filename in args.config.split(','):
+            config_filename: str
+            update_dict = None
+            if config_filename.lower().endswith('.json'):
+                update_dict = json.load(open(config_filename))
+            elif config_filename.lower().endswith('.yaml') or config_filename.lower().endswith('.yml'):
+                update_dict = yaml.load(open(config_filename))
+            elif config_filename.lower().endswith('.toml'):
+                update_dict = tomli.load(open(config_filename, 'rb'))
+            else:
+                parser.error(
+                    'Config files have to end with ".json", ".yaml", ".yml" or ".toml" to detect correct parser type')
+            if update_dict is not None:
+                config = config.updated(update_dict)
+    if args.config_keys is not None:
+        # FIXME: does not handle single quote escaping
+        separator = re.compile(
+            r"""(?!\B"[^"]*),(?![^"]*"\B)""")  # comma not between double quotes
+        toml_string = '\n'.join(re.split(separator, args.config_keys))
+        update_dict = tomli.loads(toml_string)
+        config = config.updated(update_dict)
+    if args.verbosity >= 1:
+        print(f'g_config=Config({config})')
 
     queue = Queue()
     server_process = Process(target=run_server, args=(
-        queue, args.http_host, args.http_port))
+        config, queue, args.http_host, args.http_port, args.verbosity))
     server_process.start()
 
-    print(f"""\
+    if args.verbosity >= 0:
+        print(f"""\
 Lidia GUIs driven by '{args.source}' source served on:
-  - RPC task: http://localhost:{args.http_port}
-  - Primary Flight Display: http://localhost:{args.http_port}/pfd
-  - Ship Approach: http://localhost:{args.http_port}/approach""")
+    - RPC task: http://localhost:{args.http_port}
+    - Primary Flight Display: http://localhost:{args.http_port}/pfd
+    - Ship Approach: http://localhost:{args.http_port}/approach""")
+
     try:
         (sources[args.source])(queue, args)
 
     except KeyboardInterrupt:
-        print('Exiting main loop')
+        if args.verbosity >= 0:
+            print('Exiting main loop')
 
     server_process.terminate()