+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+## [0.4.0] - 2022-12-16
+
 ### Added
 
 - Configuration with nested dictionaries using JSON, YAML and TOML files or command line arguments
@@ -20,6 +22,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Removed
 
 - Removed default values from `AircraftState`
+
+### Fixed
+
+- Altitude tape showing full units #1
 
 ## [0.3.5] - 2022-12-09
 
@@ -85,7 +91,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Initial release with `demo` and `rpctask` sources
 
 <!-- prettier-ignore -->
-[Unreleased]: https://gitlab.com/Maarrk/lidia/-/compare/v0.3.5...dev
+[Unreleased]: https://gitlab.com/Maarrk/lidia/-/compare/v0.4.0...dev
+[0.4.0]: https://gitlab.com/Maarrk/lidia/-/compare/v0.3.5...v0.4.0
 [0.3.5]: https://gitlab.com/Maarrk/lidia/-/compare/v0.3.4...v0.3.5
 [0.3.4]: https://gitlab.com/Maarrk/lidia/-/compare/v0.3.3...v0.3.4
 [0.3.3]: https://gitlab.com/Maarrk/lidia/-/compare/v0.3.2...v0.3.3